Revert "refactor(apps/desktop/AStar.Dev.OneDrive.Sync.Client): extract SyncPa…"
This reverts commit ae488de9681b5b91dbc46ebce7eb942d3b84910b.
